// src/main/webapp/app/shared/model/enumerations/organization-file-type.enum.ts
export enum OrganizationFileType {
  FINANCE_MANAGER_SIGNATURE = 'FINANCE_MANAGER_SIGNATURE',
  LOGO = 'LOGO',
  DOCUMENT_LETTER_HEAD = 'DOCUMENT_LETTER_HEAD',
  ORGANIZATION_STAMP = 'ORGANIZATION_STAMP',
  PF_STATEMENT_LETTER_HEAD = 'PF_STATEMENT_LETTER_HEAD',
  TAX_STATEMENT_LETTER_HEAD = 'TAX_STATEMENT_LETTER_HEAD',
  NOMINEE_LETTER_HEAD = 'NOMINEE_LETTER_HEAD',
  SALARY_PAYSLIP_LETTER_HEAD = 'SALARY_PAYSLIP_LETTER_HEAD',
  FESTIVAL_BONUS_PAYSLIP_LETTER_HEAD = 'FESTIVAL_BONUS_PAYSLIP_LETTER_HEAD',
  RECRUITMENT_REQUISITION_LETTER_HEAD = 'RECRUITMENT_REQUISITION_LETTER_HEAD',
}